+++ src/libs/utils.ts
@@ -25,6 +25,8 @@ export const shortFormatDate = (dateString: string): string => {
     return shortDateFormatter.format(date);
 };
 
+export const timestampNow = (): string => new Date().toISOString();
+
 export const getDiffDates = (start: string | 'now', end: string | 'now') => {
     const startDate = start === 'now' ? new Date() : new Date(start);
     const endDate = end === 'now' ? new Date() : new Date(end);