+++ src/modules/newsFeed/NewsFeedUtils.ts
@@ -10,3 +10,6 @@ const ASSET_KIND_LABELS: Record<NewsItemAssetKind, string> = {
 export const getAssetKindLabel = (assetKind: NewsItemAssetKind): string => {
     return ASSET_KIND_LABELS[assetKind];
 };
+
+// Interval to actualize unseen news count, important news & refresh banner
+export const ACTUALIZE_NEWS_INTERVAL = 60000; // 1min