+++ src/routes/computePlanDetails/ComputePlanTasks.tsx
@@ -40,14 +40,10 @@ import TasksBreadcrumbs from './components/TasksBreadCrumbs';
 interface GenericTasksProps {
     tasksTable: React.ReactNode;
     taskDrawer: React.ReactNode;
-    onRefresh: () => void;
-    tasksLoading: boolean;
 }
 const GenericTasks = ({
     tasksTable,
     taskDrawer,
-    onRefresh,
-    tasksLoading,
 }: GenericTasksProps): JSX.Element => {
     const [, params] = useRoute(ROUTES.COMPUTE_PLAN_TASKS.path);
     const key = params?.key;
@@ -82,12 +78,7 @@ const GenericTasks = ({
             >
                 <HStack justifyContent="space-between">
                     <TasksBreadcrumbs />
-                    <Actions
-                        computePlan={computePlan}
-                        loading={loading}
-                        onRefresh={onRefresh}
-                        tasksLoading={tasksLoading}
-                    />
+                    <Actions computePlan={computePlan} loading={loading} />
                 </HStack>
                 <TabsNav />
             </Box>
@@ -123,7 +114,6 @@ const TestTasks = ({
     compileListPath,
     compileDetailsPath,
 }: TasksProps): JSX.Element => {
-    const dispatchWithAutoAbort = useDispatchWithAutoAbort();
     const [page] = usePage();
     const [match] = useMatch();
     const [ordering] = useOrdering('-rank');
@@ -178,28 +168,8 @@ const TestTasks = ({
         (state) => state.computePlans.computePlan
     );
 
-    const onRefresh = () => {
-        dispatchWithAutoAbort(
-            retrieveComputePlanTestTasks({
-                computePlanKey,
-                page,
-                ordering,
-                match,
-                status,
-                worker,
-                creation_date_after: creationDateAfter,
-                creation_date_before: creationDateBefore,
-                start_date_after: startDateAfter,
-                start_date_before: startDateBefore,
-                end_date_after: endDateAfter,
-                end_date_before: endDateBefore,
-            })
-        );
-    };
-
     return (
         <GenericTasks
-            tasksLoading={loading}
             tasksTable={
                 <TasksTable
                     loading={loading}
@@ -225,7 +195,6 @@ const TestTasks = ({
                     setPageTitle={true}
                 />
             }
-            onRefresh={onRefresh}
         />
     );
 };
@@ -236,7 +205,6 @@ const TrainTasks = ({
     compileListPath,
     compileDetailsPath,
 }: TasksProps): JSX.Element => {
-    const dispatchWithAutoAbort = useDispatchWithAutoAbort();
     const [page] = usePage();
     const [match] = useMatch();
     const [ordering] = useOrdering('-rank');
@@ -291,28 +259,8 @@ const TrainTasks = ({
         (state) => state.computePlans.computePlan
     );
 
-    const onRefresh = () => {
-        dispatchWithAutoAbort(
-            retrieveComputePlanTrainTasks({
-                computePlanKey,
-                page,
-                ordering,
-                match,
-                status,
-                worker,
-                creation_date_after: creationDateAfter,
-                creation_date_before: creationDateBefore,
-                start_date_after: startDateAfter,
-                start_date_before: startDateBefore,
-                end_date_after: endDateAfter,
-                end_date_before: endDateBefore,
-            })
-        );
-    };
-
     return (
         <GenericTasks
-            tasksLoading={loading}
             tasksTable={
                 <TasksTable
                     loading={loading}
@@ -338,7 +286,6 @@ const TrainTasks = ({
                     setPageTitle={true}
                 />
             }
-            onRefresh={onRefresh}
         />
     );
 };
@@ -348,7 +295,6 @@ const CompositeTasks = ({
     compileListPath,
     compileDetailsPath,
 }: TasksProps): JSX.Element => {
-    const dispatchWithAutoAbort = useDispatchWithAutoAbort();
     const [page] = usePage();
     const [match] = useMatch();
     const [ordering] = useOrdering('-rank');
@@ -403,28 +349,8 @@ const CompositeTasks = ({
         (state) => state.computePlans.computePlan
     );
 
-    const onRefresh = () => {
-        dispatchWithAutoAbort(
-            retrieveComputePlanCompositeTasks({
-                computePlanKey,
-                page,
-                ordering,
-                match,
-                status,
-                worker,
-                creation_date_after: creationDateAfter,
-                creation_date_before: creationDateBefore,
-                start_date_after: startDateAfter,
-                start_date_before: startDateBefore,
-                end_date_after: endDateAfter,
-                end_date_before: endDateBefore,
-            })
-        );
-    };
-
     return (
         <GenericTasks
-            tasksLoading={loading}
             tasksTable={
                 <TasksTable
                     loading={loading}
@@ -450,7 +376,6 @@ const CompositeTasks = ({
                     setPageTitle={true}
                 />
             }
-            onRefresh={onRefresh}
         />
     );
 };
@@ -460,7 +385,6 @@ const AggregateTasks = ({
     compileListPath,
     compileDetailsPath,
 }: TasksProps): JSX.Element => {
-    const dispatchWithAutoAbort = useDispatchWithAutoAbort();
     const [page] = usePage();
     const [match] = useMatch();
     const [ordering] = useOrdering('-rank');
@@ -515,28 +439,8 @@ const AggregateTasks = ({
         (state) => state.computePlans.computePlan
     );
 
-    const onRefresh = () => {
-        dispatchWithAutoAbort(
-            retrieveComputePlanAggregateTasks({
-                computePlanKey,
-                page,
-                ordering,
-                match,
-                status,
-                worker,
-                creation_date_after: creationDateAfter,
-                creation_date_before: creationDateBefore,
-                start_date_after: startDateAfter,
-                start_date_before: startDateBefore,
-                end_date_after: endDateAfter,
-                end_date_before: endDateBefore,
-            })
-        );
-    };
-
     return (
         <GenericTasks
-            tasksLoading={loading}
             tasksTable={
                 <TasksTable
                     loading={loading}
@@ -562,7 +466,6 @@ const AggregateTasks = ({
                     setPageTitle={true}
                 />
             }
-            onRefresh={onRefresh}
         />
     );
 };